bentobox: small refactoring

+++ base/usr/include/kernel/vfs.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 #define MAX_PATH 256
@@ -16,14 +17,14 @@ typedef struct vfs_node {
     char name[MAX_PATH];
     bool open;
     enum vfs_node_type type;
-    uint32_t size;
+    size_t size;
     uint16_t perms;
     uint32_t inode;
     struct vfs_node *parent;
     struct vfs_node *children;
     struct vfs_node *next;
-    int32_t(*read)(struct vfs_node *node, void *buffer, uint32_t offset, uint32_t len);
-    int32_t(*write)(struct vfs_node *node, void *buffer, uint32_t offset, uint32_t len);
+    long(*read)(struct vfs_node *node, void *buffer, long offset, size_t len);
+    long(*write)(struct vfs_node *node, void *buffer, long offset, size_t len);
 } vfs_node_t;
 
 extern struct vfs_node *vfs_root;
