+++ kernel/arch/x86_64/serial.c
@@ -75,7 +75,7 @@ int dprintf(const char *fmt, ...) {
     return ret;
 }
 
-int32_t serial_write(struct vfs_node *node, void *buffer, uint32_t offset, uint32_t len) {
+long serial_write(struct vfs_node *node, void *buffer, long offset, size_t len) {
     char *buf = (char *)buffer;
     for (uint32_t i = offset; i < len; i++) {
         serial_write_char(buf[i]);
@@ -83,7 +83,7 @@ int32_t serial_write(struct vfs_node *node, void *buffer, uint32_t offset, uint3
     return (int32_t)len;
 }
 
-int32_t serial_read(struct vfs_node *node, void *buffer, uint32_t offset, uint32_t len) {
+long serial_read(struct vfs_node *node, void *buffer, long offset, size_t len) {
     // TODO: handle like in ps2.c
     char c = serial_read_char();
     memcpy(buffer, &c, 1);
